add new classifier

+++ attachment_classifier/attachment_dataset.py
@@ -28,7 +28,7 @@ class AttachmentDataset(torch.utils.data.Dataset):
         attachment = np.load(os.path.join(self.root, filename))
         data[0:len(attachment)] = attachment
 
-        data = torch.from_numpy(data)
+        data = torch.from_numpy(data, dtype=torch.float)
         #data = normalize(data)
 
         tlabel = torch.zeros(self.num_classes, dtype=torch.float)
